Added new -x flag to prevent script execution
dhcp6c will launch th script  specified in the config files on receipt
of a reply from the dhcp6 server. In certain circumstances this is not
requred, namely where rtsold runs after dhcp6c ( certain ISP's require
this ) and calls the same script. This additional flag prevents dhcp6c
from running the script only when it receives a REPLY signal from the
server. As the REPLY signal is only sent when replying to a REQUEST, and
as the REQUEST is only sent at the start of a session then any other
trigger to run the script will execute it as normal.